チルノのCOCK/requirements.txt
pygame==2.5.2
matplotlib==3.8.0
Pillow==10.0.1
